Allow any Accept form to work with accept(s)
Updates the dependencies to include newer versions of both snap and
http-media.

Both MediaType and Language values can be matched with accept and
accepts, using a new type class that defines the request and response
header names for each one. Other accept forms could be included in here
in the future.