Correct accept function, specialise function types
The change to the accept function caused it to run the given action
regardless of whether or not the relevant header matched. This change
replaces the action with 'pass'.

Generalising the two functions to work on any instance of FromHeader
meant that it was much more difficult to work with OverloadedStrings,
since both MediaType and Language are instances of IsString. Now
functions that specialise the input type to those two types are also
provided, and the two types are exported so the user does not need to
require http-media to explicitly specify the types.