Update README to use preferred style
Not only were the types of the overloaded strings in the example
ambiguous, matching on individual media types at a time should not be
encouraged, since many might be appropriate and grouping them into a
single 'accepts' call ensures that the server can pick the most
desirable one.

This change also updates the year in the license and adds standard
Haskell ignore patterns.